jblog 평과 과제 #3
- 카테고리 (미분류) 삭제 금지 (해결)

+++ jblog03/src/main/java/jblog/service/CategoryService.java
@@ -28,9 +28,13 @@ public class CategoryService {
     }
 
     @Transactional
-    public void deleteCategory(int categoryId) {
+    public boolean deleteCategory(int categoryId) {
+        if (categoryRepository.existCategoryById(categoryId)) {
+            return false;
+        }
         postRepository.deletePostById(categoryId);
         categoryRepository.deleteCategoryById(categoryId);
+        return true;
     }
 
     public void insertCategory(String id, String categoryName, String description) {
